Desafío incorporación librerías

+++ SCRIPTS/functions.js
@@ -19,6 +19,18 @@ const agregarACarrito = producto => {
   localStorage.setItem("carritoStorage", JSON.stringify(carritoStorage));
   numeroCarrito.innerHTML = JSON.parse(localStorage.getItem("numeroCarritoStorage"));
   renderizarCarrito();
+  Toastify({
+    text: "Agregado al carrito",
+    duration: 2000,
+    close: true,
+    gravity: "bottom",
+    position: "left", 
+    stopOnFocus: true,
+    style: {
+      background: "linear-gradient(to right, #00b09b, #96c93d)",
+    },
+    onClick: function() {}
+  }).showToast();
 }
 
 /* Renderizar Carrito */
@@ -45,16 +57,51 @@ const renderizarCarrito = () => {
     `
     contenidoCarrito.append(article);
   })
+  if (parseInt(numeroCarrito.innerHTML) != 0) {
+    botonVaciarCarrito.style.display = "inline";
+  }
   let totalCompra = document.createElement("div");
-  let resultado = carritoStorage.reduce((acc,elem) => acc + elem.subtotal,0);
+  let resultado = carritoStorage.reduce((acc, elem) => acc + elem.subtotal, 0);
   totalCompra.innerHTML = `<h5 class="fw-bold text-center">Total: US$${resultado} </h5>`
   contenidoCarrito.append(totalCompra);
 }
 
 /* Vaciar carrito */
 const vaciarCarrito = () => {
-  localStorage.removeItem("carritoStorage");
-  localStorage.setItem("numeroCarritoStorage", 0)
-  contenidoCarrito.innerHTML = `<h3>Su carrito está vacío</h3>`
-  numeroCarrito.innerHTML = 0;
+  Swal.fire({
+    title: 'Desea vaciar su carrito?',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Sí, vaciarlo',
+    cancelButtonText: 'Cancelar',
+    allowOutsideClick: () => {
+      const popup = Swal.getPopup()
+      popup.classList.remove('swal2-show')
+      setTimeout(() => {
+        popup.classList.add('animate__animated', 'animate__headShake')
+      })
+      setTimeout(() => {
+        popup.classList.remove('animate__animated', 'animate__headShake')
+      }, 500)
+      return false
+    }
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire(
+        '',
+        'Su carrito ha sido vaciado',
+        'success'
+      )
+      localStorage.removeItem("carritoStorage");
+      localStorage.setItem("numeroCarritoStorage", 0)
+      contenidoCarrito.innerHTML = `<h3>Su carrito está vacío</h3>`
+      carritoStorage = [];
+      numeroCarrito.innerHTML = 0;
+      botonVaciarCarrito.style.display = "none";
+    }
+  })
+
 }
+
